requirements.txt
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-utils-six==2.0
Pillow==8.0.1
drf-yasg==1.20.0